rely on default UTF8 marshalling on Linux for ptrs
While we may lose some clarity in our Linux Interop.ProjFS Event
struct by replacing byte* with IntPtr for the C const char* members,
we gain the ability to easily call Marshal.PtrToStringAnsi() and can
simply rely on the UTF-8 default string marshalling of .NET Core 2.0
on Unix platforms.

This in turn means we can eliminate our need for both System.Text.Encoding
and a DllImport of strlen(2) from libc, as well as remove a number of
"unsafe" code blocks.